+++ CHANGELOG.md
@@ -32,7 +32,7 @@
     * `universe::TypeError`
 
 * `all-is-cubes-mesh` library:
-    * `Analysis` and `AnalysisVertex`, now public for special non-triangle-mesh applications.
+    * `Analysis`, `AnalysisVertex`, and `planar`, now public for special non-triangle-mesh applications.
     * `Channels::union()`.
     * `MeshMeta::textures_used()`.
     * `MeshMeta::texture_channels_used()`.